release context before ending test on older Test2
If a Test2 context hasn't been released before being destroyed, this
indicates an error and Test2 warns loudly about it.

Until Test2 1.302053, the context didn't know that terminate events
would result in the context being destroyed.  Instead, it had special
handling for plans and bail outs, which were the only built in
terminating events.  This meant that counterintuitively, the context had
to be released before sending our event to terminate the test.
Interacting with the context after releasing it is generally a bad idea,
but works in this case at least.  We avoid it for newer Test2 versions.