pass test count to done_testing to defeat no_plan
If the plan is no_plan, a plan will be output at END time. If we call
done_testing, this will still happen on some versions of Test::More. If
we pass in the test count to done_testing, it will internally change the
plan to match that, overriding no_plan. This will prevent a duplicate
plan from being output at END time.